requirements.txt
gradio
azure-core
azure-ai-textanalytics
